Start of spiped role
Split Debian/FreeBSD task files for installation and configuration of
spiped for tunneling node_exporter traffic back to the main host.

Uses git-crypt to encrypt the keyfile which is versioned in the
repository to prevent having to distribute it out of band.

Completely untested as yet. Binds the tunnel on each host to 9101 which
is targeted at the 9100 port that node_exporter listens on privately.